fix(users): Fix wring field type

+++ app/Modules/Users/Database/Migrations/0000_00_00_100000_create_users_tables.php
@@ -19,7 +19,7 @@ class CreateUsersTables extends Migration
 			{
 				$table->increments('id');
 				$table->string('name', 255);
-				$table->tinyInteger('puid')->unsigned()->default(0);
+				$table->integer('puid')->unsigned()->default(0);
 				$table->index('name');
 				$table->index('puid');
 			});
